pyproject.toml
[project]
name = "openai-agent-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "cryptography>=44.0.2",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "httpx>=0.28.1",
    "markdown>=3.8",
    "markdownify>=1.1.0",
    "mcp[cli]>=1.6.0",
    "openai>=1.76.0",
    "openai-agents>=0.0.13",
    "pandas>=2.2.3",
    "problems>=0.0.2",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "quart>=0.20.0",
    "quart-cors>=0.8.0",
    "reportlab>=4.4.0",
    "requests>=2.32.3",
    "weasyprint>=65.1",
]
